Update .gitignore, README, and various components for improved functionality and clarity
- Added `trader/runs/` to .gitignore to prevent tracking of temporary trader run files.
- Clarified README documentation regarding the trader's local prediction list and its feedback emission behavior.
- Enhanced `Maker` component to ensure adverse selection penalties are calculated correctly.
- Refactored `CausalSymbol` to improve sample management and maintain history cap.
- Updated configuration tests for better validation of default values.
- Improved error handling in measurement logic to log invalid timeframes.
- Streamlined prediction handling in the `Crypto` component by removing redundant feedback logging.
- Added tests for lead-lag dominance calculations to ensure accurate margin handling.

+++ broker/maker.go
@@ -59,7 +59,7 @@ func (maker *Maker) FillPaper(tradingWallet *wallet.Wallet) (order.Fill, error) 
 
 	orderSymbol := Symbol(maker.Symbol)
 	base := orderSymbol.BaseAsset()
-	effectivePrice := maker.LimitPrice * (1 + config.System.AdverseSelectionBPS/10000)
+	effectivePrice := maker.LimitPrice * (1 + float64(config.System.AdverseSelectionBPS)/10000)
 	qty := (maker.Notional - fee) / effectivePrice
 
 	if qty <= 0 {
